Update objhunt.txt
This is just a suggestion. I would like to add the de_* maps to Obj Hunt in the Singleplayer Menu like in the original Prop Hunt Gamemode. And more importantly I would like to add the workshop id of the addon. This makes the server telling the client to download the Gamemode's Workshop addon. In addition you don't need to download the taunts through the server anymore that are already in the gamemode. (Maybe just let the server search for taunts in the garrysmod root directory?)